Enhance CURSOR-GET to support :SET, :SET-KEY, and :SET-RANGE.
Add a test to the test-suite for :SET-RANGE.

Fix a bug in CLOSE-CURSOR, it was calling LIBLMDB:CURSOR-CLOSE on the handle, not its contents.

Make each of the test suite tests initialize its data. They are run in basically random order, so none can depend on any others.

(lmdb:run-tests) now works as expected.